# openapi.yml
openapi: 3.0.3
info:
  title: PrimeraApiCasa - Biblioteca API
  description: |
    REST API para gestión de una biblioteca. Usa autenticación JWT.

    **Roles:**
    - `USER` — puede consultar el catálogo (autores, libros, detalles, categorías, libro-categorías)
    - `ADMIN` — acceso completo, incluyendo escritura en el catálogo y gestión de roles y usuarios

    **Autenticación:** Los endpoints protegidos requieren el header `Authorization: Bearer <token>`.
    El token se obtiene en `/auth/login` o `/auth/register`.
  version: 1.0.0
servers:
  - url: http://localhost:8080/micro
    description: Servidor local
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Registrar nuevo usuario
      description: Crea un usuario nuevo y devuelve un JWT. No requiere autenticación.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: juan123
              password: secreto123
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: El nombre de usuario ya existe
  /auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      description: Autentica al usuario y devuelve un JWT. No requiere autenticación.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: juan123
              password: secreto123
      responses:
        '200':
          description: Autenticación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciales incorrectas

  # ─── AUTORES ─────────────────────────────────────────────────────────────────
  /autores:
    get:
      tags: [Autores]
      summary: Listar todos los autores
      description: Requiere rol `USER` o `ADMIN`.
      responses:
        '200':
          description: Lista de autores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AutorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Autores]
      summary: Crear un autor
      description: Requiere rol `ADMIN`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutorCreateRequest'
            example:
              nombre: Gabriel García Márquez
              nacionalidad: Colombiana
      responses:
        '201':
          description: Autor creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /autores/{id}:
    get:
      tags: [Autores]
      summary: Obtener autor por ID
      description: Requiere rol `USER` o `ADMIN`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Autor encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    put:
      tags: [Autores]
      summary: Actualizar un autor
      description: Requiere rol `ADMIN`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutorCreateRequest'
            example:
              nombre: Gabriel García Márquez
              nacionalidad: Colombiana
      responses:
        '200':
          description: Autor actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [Autores]
      summary: Eliminar un autor
      description: Requiere rol `ADMIN`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Autor eliminado
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ─── LIBROS ──────────────────────────────────────────────────────────────────
...
  /libros/autor/{autorId}:
    get:
      tags: [Libros]
      summary: Listar libros por autor
      description: Requiere rol `USER` o `ADMIN`.
      parameters:
        - name: autorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del autor
      responses:
        '200':
          description: Lista de libros del autor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LibroResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ─── DETALLES ────────────────────────────────────────────────────────────────
...
  /categorias:
    get:
      tags: [Categorias]
      summary: Listar todas las categorías
      description: Requiere rol `USER` o `ADMIN`.
      responses:
        '200':
          description: Lista de categorías
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoriaResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Categorias]
      summary: Crear una categoría
      description: Requiere rol `ADMIN`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoriaCreateRequest'
            example:
              nombre: Novela
      responses:
        '201':
          description: Categoría creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriaResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /categorias/{id}:
    get:
      tags: [Categorias]
      summary: Obtener categoría por ID
      description: Requiere rol `USER` o `ADMIN`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Categoría encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriaResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    put:
      tags: [Categorias]
      summary: Actualizar una categoría
      description: Requiere rol `ADMIN`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoriaCreateRequest'
            example:
              nombre: Ciencia Ficción
      responses:
        '200':
          description: Categoría actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriaResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [Categorias]
      summary: Eliminar una categoría
      description: Requiere rol `ADMIN`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Categoría eliminada
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ─── LIBRO-CATEGORIAS ────────────────────────────────────────────────────────
...
  /libro-categorias/categoria/{categoriaId}:
    get:
      tags: [LibroCategorias]
      summary: Obtener libros de una categoría
      description: Requiere rol `USER` o `ADMIN`.
      parameters:
        - name: categoriaId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID de la categoría
      responses:
        '200':
          description: Lista de asociaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LibroCategoriaResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /usuario-roles/rol/{rolId}:
    get:
      tags: [UsuarioRoles]
      summary: Obtener usuarios con un rol
      description: Requiere rol `ADMIN`.
      parameters:
        - name: rolId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del rol
      responses:
        '200':
          description: Lista de asignaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioRolResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: ID del recurso
  responses:
    Unauthorized:
      description: Token ausente o inválido
    NotFound:
      description: Recurso no encontrado
    Forbidden:
      description: No tienes permisos para este recurso
  schemas:
...
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: juan123
        password:
          type: string
          example: secreto123
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: juan123
        password:
          type: string
          example: secreto123
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
        username:
          type: string
          example: juan123
        roles:
          type: array
          items:
            type: string
          example: [ROLE_USER]

    # Autores
    AutorCreateRequest:
      type: object
      required: [nombre]
      properties:
        nombre:
          type: string
          maxLength: 100
          example: Gabriel García Márquez
        nacionalidad:
          type: string
          maxLength: 50
          example: Colombiana
    AutorResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nombre:
          type: string
          example: Gabriel García Márquez
        nacionalidad:
          type: string
          example: Colombiana
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Libros
...
    LibroResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        titulo:
          type: string
          example: Cien años de soledad
        anioPublicacion:
          type: integer
          example: 1967
        autorId:
          type: integer
          format: int64
          example: 1
        autorNombre:
          type: string
          example: Gabriel García Márquez
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Detalles
...
    CategoriaCreateRequest:
      type: object
      required: [nombre]
      properties:
        nombre:
          type: string
          maxLength: 100
          example: Novela
    CategoriaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nombre:
          type: string
          example: Novela

    # LibroCategorias
...
    LibroCategoriaResponse:
      type: object
      properties:
        libroId:
          type: integer
          format: int64
          example: 1
        libroTitulo:
          type: string
          example: Cien años de soledad
        categoriaId:
          type: integer
          format: int64
          example: 2
        categoriaNombre:
          type: string
          example: Novela
        prioridad:
          type: integer
          example: 1
        addedAt:
          type: string
          format: date-time
        comentario:
          type: string
          example: Categoría principal

    # Roles
...
    UsuarioRolResponse:
      type: object
      properties:
        usuarioId:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: juan123
        rolId:
          type: integer
          format: int64
          example: 2
        rolNombre:
          type: string
          example: ADMIN